moved each file to there respective folder
