+++ console/console-init/ui/src/modules/address-detail/components/AddressLinks/AddressLinks.tsx
