+++ console/console-init/ui/src/modules/address-detail/components/TypeBadge/TypeBadge.tsx
