+++ console/console-init/ui/src/modules/address-detail/components/index.ts
@@ -1,3 +1,9 @@
+/*
+ * Copyright 2020, EnMasse authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+
 export * from "./AddressDetailHeader";
 export * from "./AddressLinks";
+export * from "./TypeBadge";
 export * from "./EmptyAddressLinks";
