+++ console/console-init/ui/src/modules/address/components/AddressList/index.ts
@@ -3,6 +3,4 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-export * from "./AddressListContainer";
-export * from "./AddressListFilter";
-export * from "./AddressListToolbar";
+export * from "./AddressList";
