+++ console/console-init/ui/src/modules/address/components/EmptyAddress/index.ts
@@ -4,6 +4,3 @@
  */
 
 export * from "./EmptyAddress";
-export * from "./AddressList";
-export * from "./Messages";
-export * from "./TypePlan";
