+++ console/console-init/ui/src/modules/address/containers/AddressListContainer/index.ts
@@ -4,5 +4,3 @@
  */
 
 export * from "./AddressListContainer";
-export * from "./AddressListFilter";
-export * from "./AddressListToolbar";
