+++ console/console-init/ui/src/modules/address/dialogs/index.ts
@@ -3,6 +3,5 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-export * from "./AddressListContainer";
-export * from "./AddressListFilter";
-export * from "./AddressListToolbar";
+export * from "./CreateAddress";
+export * from "./EditAddress";
