+++ console/console-init/ui/src/stories/AddressDetailHeader.stories.tsx
@@ -7,7 +7,7 @@ import React from "react";
 import { MemoryRouter } from "react-router";
 import { text, number, select } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
-import { AddressDetailHeader } from "modules/address-detail/components/AddressDetailHeader";
+import { AddressDetailHeader } from "modules/address-detail/components/AddressDetailHeader/AddressDetailHeader";
 
 export default {
   title: "Address Detail"
